pyproject.toml
[project]
name = "certif-gcp-mle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dvc>=3.67.1",
    "pdfplumber>=0.11.9",
    "pyyaml>=6.0",
]
